ticksignals-mcp/requirements.txt
mcp>=1.20.0
httpx>=0.27.0
yfinance>=0.2.30
pandas>=2.0.0
numpy>=1.24.0
prophet>=1.1.5
python-dotenv>=1.0.0
